requirements.txt
astroid==2.3.1
Click==7.0
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack-python==0.5.6
pykeybase==0.4.1
pylint==2.4.2
python-dotenv==0.10.3
six==1.12.0
SQLAlchemy==1.3.8
typed-ast==1.4.0
uWSGI==2.0.18
Werkzeug==0.16.0
wrapt==1.11.2
